Rename `amx` to `xr`, `xr/wua` to `xr/lua`
