+++ gamedata/scripts/xr/registrator.lua
